requirements.txt
# 使用 PyPI + 額外的 PyTorch CUDA 來源
# 這樣 torch / torchvision 會從 CUDA 12.1 的索引抓，其他套件仍從 PyPI 抓
--extra-index-url https://download.pytorch.org/whl/cu121

# === PyTorch (GPU) ===
torch==2.4.0
torchvision==0.19.0

# === Hugging Face / Diffusers 生態系 ===
transformers==4.48.3
diffusers==0.30.0
accelerate==0.33.0
safetensors==0.4.5
huggingface-hub==0.25.0
datasets>=2.20.0
pyarrow>=15.0.0

# === 影像與數值處理 ===
Pillow==10.4.0
numpy==1.26.4

# === 其他常見相依，列出來比較穩 ===
requests==2.32.3
tqdm==4.66.5
protobuf==5.28.3
